refactor: 선택된 포켓몬의 slice 명 변경
#17
